feat(business/tabs): separate business and public component-viewer into core parts
