--- projects/angular-showcase/src/app/features/business-component-viewer/business-component-viewer/business-component-viewer.component.html
@@ -1,19 +1,0 @@
-<perfect-scrollbar>
-  <article>
-    <sbb-tabs>
-      <sbb-tab label="Overview">
-        <div #overview></div>
-      </sbb-tab>
-      <sbb-tab label="API">
-        <div #api></div>
-      </sbb-tab>
-      <sbb-tab *ngIf="example | async as examplePortals" label="Examples">
-        <sbb-business-example-viewer
-          *ngFor="let entry of examplePortals | keyvalue"
-          [name]="entry.key"
-          [example]="entry.value"
-        ></sbb-business-example-viewer>
-      </sbb-tab>
-    </sbb-tabs>
-  </article>
-</perfect-scrollbar>
