+++ projects/angular-showcase/src/app/features/core-component-viewer/core-component-viewer/core-component-viewer.component.scss
