+++ projects/angular-showcase/src/app/features/core-component-viewer/core-example-viewer/core-example-viewer.component.html
