+++ projects/angular-showcase/src/app/features/ps-chrome-patch/ps-chrome-patch.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+
+import { PsChromePatchDirective } from './ps-chrome-patch.directive';
+
+@NgModule({
+  declarations: [PsChromePatchDirective],
+  exports: [PsChromePatchDirective]
+})
+export class PsChromePatchModule {}
