--- projects/angular-showcase/src/app/features/public-component-viewer/public-component-viewer/public-component-viewer.component.scss
@@ -1,12 +1,0 @@
-:host {
-  display: block;
-  margin-left: 300px;
-}
-
-perfect-scrollbar {
-  height: calc(100vh - 48px);
-}
-
-article {
-  padding: 3rem;
-}
