+++ projects/angular-showcase/src/app/features/public-component-viewer/public-component-viewer/public-component-viewer.component.ts
@@ -6,9 +6,12 @@ import { ExampleProvider } from '../../../shared/example-provider';
 import { HtmlLoader } from '../../../shared/html-loader.service';
 
 @Component({
-  selector: 'sbb-public-component-viewer',
-  templateUrl: './public-component-viewer.component.html',
-  styleUrls: ['./public-component-viewer.component.scss']
+  selector: 'sbb-component-viewer',
+  templateUrl:
+    '../../core-component-viewer/core-component-viewer/core-component-viewer.component.html',
+  styleUrls: [
+    '../../core-component-viewer/core-component-viewer/core-component-viewer.component.scss'
+  ]
 })
 export class PublicComponentViewerComponent extends ComponentViewerBase {
   constructor(
